requirements.txt
Flask==2.1.2
flask-cors==3.0.10
openai==0.27.0
python-dotenv==1.0.0
